fix(perf): direct-emit review round — sizing pin de-vacuated, sectionY parity, width guard, diag receipt
One-lens Opus review, no correctness defect (record in v0.10.0-progress.md).
The exactPreSizing pin had gone vacuous (all-transcoded columns route direct
past the sizing arithmetic) — both twins now serialize a MIXED column with a
routing premise. Out-of-byte-range sectionY now truncates identically to the
native route's writeByte (the one demonstrated identity gap; corpus/tool
overload only). convertIndexed enforces the native indexed-width thresholds
instead of assuming the transcoder's fallback gate, with translator-suite
identity + width-guard pins. The fuzz asserts a nonzero direct fraction; the
routing pin's negative arms assert servable bytes; direct_v20= joins the
Fabric read_path diag line as the live liveness receipt.

+++ fabric/src/main/java/dev/vox/lss/networking/server/NbtSectionSerializer.java
@@ -974,7 +974,13 @@ final class NbtSectionSerializer {
         for (var p : parsed) {
             var t = p.transcoded();
             sections.add(new WireSectionCursor.WireSection(
-                    p.sectionY(), p.nonEmptyCount(), p.fluidCount(),
+                    // (byte) cast: the native route's writeByte TRUNCATES an
+                    // out-of-byte-range sectionY (translate then round-trips the
+                    // truncated value), so the direct route must truncate identically —
+                    // production gates Y to the world range, but the range-free corpus/
+                    // tool overload serializes garbage Y and the byte-identity claim
+                    // must hold there too (review finding 2).
+                    (byte) p.sectionY(), p.nonEmptyCount(), p.fluidCount(),
                     dev.vox.lss.common.wire.NativeToV20Translator.convertIndexed(
                             t.blockBits(), t.blockIds(), t.blockData(), true, dict, blockIdentity),
                     dev.vox.lss.common.wire.NativeToV20Translator.convertIndexed(
